requirements.txt
fastapi
uvicorn
replicate
python-multipart
jinja2
python-dotenv
